Make the system more configurable and support source directory
Always extracting to a temporary directory for source tarballs allows to
always be in the directory use-case, be it from a tarball or from a
source directory.

Framing with RPM as a context manager for tempdir cleanup.

Move more parameters to the (inherited) template sections rather than
the pysrpm section. Also re-evaluate inherited parent templates in the child
sections, so that section-local references e.g. ${package} are updated.

Re-using a parent template requires specifying the section, e.g. in
fedora-pre8 you could use ${fedora:...} or ${base:...}